split polling_policy.h into command spanner portions
