+++ google/cloud/CMakeLists.txt
@@ -153,6 +153,7 @@ add_library(
     log.cc
     log.h
     optional.h
+    polling_policy.h
     status.cc
     status.h
     status_or.h
